# contracts/api.yaml
...
openapi: 3.0.3
info:
  title: "AML Case Management API"
  version: "1.0.0"
  description: |
    API contracts for the AML Case Management System.

    ## Overview

    This API enables Spriggy to meet AUSTRAC AML/CTF compliance obligations by providing
    endpoints for managing KYC remediation, PEP/Sanctions screening disposition, and
    suspicious activity investigation workflows.

    ## Authentication

    All endpoints require Bearer token authentication via OIDC SSO integration (D5).
    Include the token in the Authorization header:

    ```
    Authorization: Bearer <access_token>
    ```

    Webhook endpoints use HMAC signature validation instead of Bearer tokens (D11).

    ## Roles

    The system enforces four distinct roles (FR-063):
    - **L1_ANALYST**: Triage and basic case handling
    - **L2_ANALYST**: Full investigation and SMR recommendation authority
    - **AML_MANAGER**: Approval authority for SMRs
    - **READ_ONLY**: View and export only

    ## Error Handling

    All errors follow a consistent format with machine-readable codes.
    See the ErrorResponse schema for details.

    ## Optimistic Locking

    Case mutations require a `version` field for optimistic locking (D8).
    Version mismatch returns 409 Conflict.

    ## Constitution Alignment

    - **Principle I**: All mutating endpoints create audit log entries
    - **Principle II**: Endpoints enforce role-based access
...
  contact:
    name: AML Compliance Team
...
servers:
  - url: /api/v1
...
paths:
...
  /cases:
    get:
      summary: List cases
      description: |
        Retrieve paginated list of cases with filtering.
        Default ordering: SLA urgency (most urgent first), then creation date (oldest first).

        **Roles**: All authenticated users

        **Source Requirements**: FR-024, FR-025, US-12
      operationId: listCases
      tags:
        - Cases
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          description: Filter by case status
          schema:
            $ref: '#/components/schemas/CaseStatus'
        - name: caseType
          in: query
          description: Filter by case type
          schema:
            $ref: '#/components/schemas/CaseType'
        - name: assignedUserId
          in: query
          description: Filter by assigned analyst (use 'me' for current user, 'unassigned' for queue)
          schema:
            type: string
        - name: tier
          in: query
          description: Filter by analyst tier
          schema:
            $ref: '#/components/schemas/Tier'
        - name: slaStatus
          in: query
          description: Filter by SLA status
          schema:
            type: string
            enum: [on_track, warning, breached]
        - name: customerId
          in: query
          description: Filter by customer ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of cases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /cases/{caseId}/edd-checklist:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    get:
      summary: Get EDD checklist
      description: |
        Retrieve EDD checklist for a PEP case.

        **Roles**: All authenticated users

        **Source Requirements**: FR-033, US-7
      operationId: getEDDChecklist
      tags:
        - EDD
      responses:
        '200':
          description: EDD checklist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EDDChecklistResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      summary: Submit EDD checklist
      description: |
        Submit completed EDD checklist for a PEP case.

        **Side Effects**:
        - Clears onboarding block via Spriggy API callback (FR-028)
        - Creates audit trail entry

        **Roles**: L1_ANALYST, L2_ANALYST, AML_MANAGER

        **Source Requirements**: FR-033, FR-034, US-7
      operationId: submitEDDChecklist
      tags:
        - EDD
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitEDDChecklistRequest'
      responses:
        '201':
          description: EDD checklist submitted, block cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EDDChecklistResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: EDD checklist already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Case is not a PEP-EDD-REQUIRED case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --------------------------------------------------------------------------
  # Onboarding Blocks
  # --------------------------------------------------------------------------
...
  /config/holidays:
    get:
      summary: List holiday overrides
      description: |
        Retrieve custom holiday dates for SLA calculation.

        **Roles**: All authenticated users

        **Source Requirements**: D2, FR-048
      operationId: listHolidayOverrides
      tags:
        - Configuration
      parameters:
        - name: year
          in: query
          description: Filter by year
          schema:
            type: integer
      responses:
        '200':
          description: Holiday overrides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayOverrideListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Add holiday override
      description: |
        Add a custom holiday date for SLA calculation.

        **Roles**: AML_MANAGER
      operationId: addHolidayOverride
      tags:
        - Configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddHolidayOverrideRequest'
      responses:
        '201':
          description: Holiday added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayOverrideResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Holiday already exists for this date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# ============================================================================
# COMPONENTS
# ============================================================================
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from OIDC SSO

  # --------------------------------------------------------------------------
  # Common Parameters
  # --------------------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CaseIdParam:
      name: caseId
      in: path
      required: true
      description: Case unique identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: One or more fields are invalid
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Authentication required
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: You do not have permission to access this resource
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: The requested resource was not found
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable message
          example: One or more fields are invalid
        details:
          type: object
          additionalProperties: true
          description: Additional error context
...
    CaseType:
      type: string
      enum:
        - KYC_REMEDIATION
        - PEP_SANCTIONS_SCREENING
        - SUSPICIOUS_ACTIVITY
        - SUPPLEMENTARY_SMR
    CaseSubtype:
      type: string
      enum:
        - SANCTIONS_ONBOARDING
        - SANCTIONS_EXISTING_CUSTOMER
        - PEP_EDD_REQUIRED
        - PEP_PROVISIONAL_REVIEW
    CaseStatus:
      type: string
      enum:
        - OPEN
        - IN_PROGRESS
        - PENDING_INFORMATION
        - ESCALATED
        - PENDING_APPROVAL
        - CLOSED
        - SMR_FILED
    L2ReviewStatus:
      type: string
      enum:
        - NOT_REQUIRED
        - PENDING_REVIEW
        - REVIEWED_ACCEPTED
        - REVIEWED_REOPENED
    Tier:
      type: string
      enum:
        - L1
        - L2
    UserRole:
      type: string
      enum:
        - L1_ANALYST
        - L2_ANALYST
        - AML_MANAGER
        - READ_ONLY
...
    EnhancedMonitoringFrequency:
      type: string
      enum:
        - WEEKLY
        - MONTHLY
        - QUARTERLY
        - ANNUALLY

    # --- Webhook Schemas ---
...
    CaseSummary:
      type: object
      required:
        - id
        - referenceCode
        - caseType
        - status
        - tier
        - slaDeadline
        - slaStatus
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        referenceCode:
          type: string
          example: AML-1001
        caseType:
          $ref: '#/components/schemas/CaseType'
        caseSubtype:
          $ref: '#/components/schemas/CaseSubtype'
        status:
          $ref: '#/components/schemas/CaseStatus'
        tier:
          $ref: '#/components/schemas/Tier'
        slaDeadline:
          type: string
          format: date-time
        slaStatus:
          type: string
          enum: [on_track, warning, breached]
        slaBreached:
          type: boolean
        l2ReviewStatus:
          $ref: '#/components/schemas/L2ReviewStatus'
        assignedUser:
          $ref: '#/components/schemas/UserSummary'
        customer:
          $ref: '#/components/schemas/CustomerSummary'
        createdAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
...
    CaseListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CaseSummary'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    CustomerSummary:
      type: object
      description: |
        Customer snapshot captured at case creation time.

        **Immutability**: Customer data is captured as a snapshot when the case is created.
        Subsequent changes to the customer's profile in Spriggy do not update this snapshot.
        This ensures case context reflects the state at the time of the alert, which is
        important for audit and investigation purposes. If updated customer information
        is needed, analysts can reference Spriggy's customer management system directly.
      required:
        - id
        - externalId
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
          description: Spriggy customer identifier
        displayName:
          type: string
          description: Masked or authorized display name
        accountStatus:
          type: string
          enum: [ACTIVE, SUSPENDED, CLOSED]
          description: Customer account status at time of case creation (snapshot)
        onboardingStatus:
          type: string
          enum: [IN_PROGRESS, COMPLETED, BLOCKED, UNKNOWN]
          description: Customer onboarding status at time of case creation (snapshot)
        snapshotAt:
          type: string
          format: date-time
          description: Timestamp when this customer data snapshot was captured

    # --- User Schemas ---
    UserSummary:
      type: object
      required:
        - id
        - displayName
        - role
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
...
    EDDChecklistResponse:
      type: object
      required:
        - id
        - caseId
        - sourceOfWealth
        - sourceOfFunds
        - relationshipPurpose
        - enhancedMonitoringFrequency
        - completedAt
        - completedBy
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        sourceOfWealth:
          type: string
        sourceOfFunds:
          type: string
        relationshipPurpose:
          type: string
        enhancedMonitoringFrequency:
          $ref: '#/components/schemas/EnhancedMonitoringFrequency'
        additionalNotes:
          type: string
        completedAt:
          type: string
          format: date-time
        completedBy:
          $ref: '#/components/schemas/UserSummary'
    SubmitEDDChecklistRequest:
      type: object
      required:
        - sourceOfWealth
        - sourceOfFunds
        - relationshipPurpose
        - enhancedMonitoringFrequency
      properties:
        sourceOfWealth:
          type: string
          minLength: 50
        sourceOfFunds:
          type: string
          minLength: 50
        relationshipPurpose:
          type: string
          minLength: 20
        enhancedMonitoringFrequency:
          $ref: '#/components/schemas/EnhancedMonitoringFrequency'
        additionalNotes:
          type: string

    # --- Onboarding Block Schemas ---
...
    HolidayOverrideResponse:
      type: object
      required:
        - id
        - holidayDate
        - description
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        holidayDate:
          type: string
          format: date
        description:
          type: string
        region:
          type: string
        createdBy:
          $ref: '#/components/schemas/UserSummary'
        createdAt:
          type: string
          format: date-time
    HolidayOverrideListResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/HolidayOverrideResponse'
    AddHolidayOverrideRequest:
      type: object
      required:
        - holidayDate
        - description
      properties:
        holidayDate:
          type: string
          format: date
        description:
          type: string
          maxLength: 200
        region:
          type: string
          maxLength: 50

    # --- Pagination ---
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
        hasNext:
          type: boolean
        hasPrev:
          type: boolean

# ============================================================================
# DEFAULT SECURITY
# ============================================================================
security:
  - bearerAuth: []